deps/v8/test/inspector/regress/regress-crbug-1283049-expected.txt

Regression test for crbug/1283049

Running test: test
foo (foo.js:0:17)
(anonymous) (:0:0)
